requirements.txt
Flask
Flask-sqlalchemy
Flask-Migrate
python-dotenv
python-slugify